Implement owner.service.spec testing and change filtering to database
Also fixed the filtering on the owner controller as it was previously checking for exact comparisons.
Added an extra owner to the database as we previously only had faculty from the science building.

filterOwners still needs to be removed, as everything is currently being filtered through the database and it serves no purpose.
Closes  #17
Progresses #16